feat: Update imports and remove legacy code
