+++ .gitignore
@@ -15,3 +15,4 @@ Utilities/Docker/*.tar.gz
 .swiftpm
 /build
 *.pyc
+/android/ExampleApp/.idea
