--- android/activitysources/src/main/java/com/fjuul/sdk/activitysources/entities/internal/HCDataManager.java
@@ -1,248 +1,0 @@
-package com.fjuul.sdk.activitysources.entities.internal;
-
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-
-import com.fjuul.sdk.activitysources.entities.FitnessMetricsType;
-import com.fjuul.sdk.activitysources.entities.HealthConnectIntradaySyncOptions;
-import com.fjuul.sdk.activitysources.entities.HealthConnectProfileSyncOptions;
-import com.fjuul.sdk.activitysources.entities.internal.healthconnect.HCCalorieDataPoint;
-import com.fjuul.sdk.activitysources.entities.internal.healthconnect.HCDataConverter;
-import com.fjuul.sdk.activitysources.entities.internal.healthconnect.HCHeartRateSummaryDataPoint;
-import com.fjuul.sdk.activitysources.entities.internal.healthconnect.HCStepsDataPoint;
-import com.fjuul.sdk.activitysources.exceptions.GoogleFitActivitySourceExceptions;
-import com.fjuul.sdk.activitysources.http.services.ActivitySourcesService;
-import com.fjuul.sdk.core.ApiClient;
-import com.fjuul.sdk.core.http.utils.ApiCallResult;
-import com.fjuul.sdk.core.utils.Logger;
-import com.google.android.gms.tasks.Task;
-import com.google.android.gms.tasks.TaskCompletionSource;
-import com.google.android.gms.tasks.Tasks;
-
-import androidx.annotation.NonNull;
-import androidx.health.connect.client.records.HeartRateRecord;
-import androidx.health.connect.client.records.HeightRecord;
-import androidx.health.connect.client.records.Record;
-import androidx.health.connect.client.records.StepsRecord;
-import androidx.health.connect.client.records.TotalCaloriesBurnedRecord;
-import androidx.health.connect.client.records.WeightRecord;
-
-public class HCDataManager {
-    private static final int MAX_DAYS = 30; // Don't try to get older data than this
-    private static final ExecutorService localBackgroundExecutor = Executors.newCachedThreadPool();
-
-    private static final String PROFILE_CHANGES_TOKEN_KEY = "hc-profile-changes-token";
-    private static final String INTRADAY_CHANGES_TOKEN_KEY = "hc-intraday-changes-token";
-
-    private final @NonNull ActivitySourcesService activitySourcesService;
-    private final @NonNull HCClientWrapper clientWrapper;
-    private final @NonNull ApiClient apiClient;
-
-    public HCDataManager(@NonNull HCClientWrapper clientWrapper,
-                         @NonNull ActivitySourcesService activitySourcesService,
-                         @NonNull ApiClient apiClient) {
-        this.clientWrapper = clientWrapper;
-        this.activitySourcesService = activitySourcesService;
-        this.apiClient = apiClient;
-    }
-
-    @NonNull
-    public Task<Void> syncIntradayMetrics(@NonNull HealthConnectIntradaySyncOptions options) {
-        Logger.get().d("Syncing intraday metrics");
-        final Task<Boolean> checkPermissionsTask = checkPermissions();
-        final Task<HealthConnectRecords> readTask =
-            checkPermissionsTask.continueWithTask(localBackgroundExecutor,
-                task -> readIntradayMetrics(options.getMetrics()));
-        return readTask.onSuccessTask(localBackgroundExecutor, (healthConnectRecords -> {
-            final String nextToken = healthConnectRecords.getNextToken();
-            final List<Record> healthRecords = healthConnectRecords.getRecords();
-            if (healthRecords.isEmpty()) {
-                Logger.get().d("No new records to send");
-                return Tasks.forResult(null);
-            } else {
-                final List<HCStepsDataPoint> stepsDataPoints = new ArrayList<>();
-                final List<HCCalorieDataPoint> calorieDataPoints = new ArrayList<>();
-                final List<HCHeartRateSummaryDataPoint> heartRateSummaryDataPoints =
-                    new ArrayList<>();
-                for (Record healthRecord : healthRecords) {
-                    if (healthRecord instanceof StepsRecord stepsRecord) {
-                        stepsDataPoints.add(HCDataConverter.convertRecordToSteps(stepsRecord));
-                    } else if (healthRecord instanceof HeartRateRecord heartRateRecord) {
-                        heartRateSummaryDataPoints.add(
-                            HCDataConverter.convertRecordToHeartRateSummary(heartRateRecord));
-                    } else if (healthRecord instanceof TotalCaloriesBurnedRecord totalCaloriesBurnedRecord) {
-                        calorieDataPoints.add(
-                            HCDataConverter.convertRecordToCalories(totalCaloriesBurnedRecord));
-                    } else {
-                        Logger.get().e("Unexpected record type: %s", healthRecord.getClass().getCanonicalName());
-                    }
-                }
-                HCUploadData uploadData = new HCUploadData();
-                uploadData.setStepsData(stepsDataPoints);
-                uploadData.setHeartRateData(heartRateSummaryDataPoints);
-                uploadData.setCaloriesData(calorieDataPoints);
-
-                return sendHCUploadData(INTRADAY_CHANGES_TOKEN_KEY, nextToken, uploadData).onSuccessTask(
-                    localBackgroundExecutor, (apiCallResult) -> Tasks.forResult(apiCallResult.getValue()));
-            }
-        }));
-    }
-
-    @NonNull
-    public Task<Void> syncProfile(@NonNull HealthConnectProfileSyncOptions options) {
-        Logger.get().d("Syncing profile");
-        final Task<Boolean> checkPermissionsTask = checkPermissions();
-        final Task<HealthConnectRecords> readTask =
-            checkPermissionsTask.continueWithTask(localBackgroundExecutor,
-                task -> readProfileMetrics(options.getMetrics()));
-        return readTask.onSuccessTask(localBackgroundExecutor, (healthConnectRecords -> {
-            final String nextToken = healthConnectRecords.getNextToken();
-            final List<Record> healthRecords = healthConnectRecords.getRecords();
-            HCSynchronizableProfileParams profileParams = new HCSynchronizableProfileParams();
-            // We have to go through all the height and weight records and select the most
-            // recent values, if any
-            Instant newestHeightTime = Instant.ofEpochMilli(0);  // very old
-            Instant newestWeightTime = Instant.ofEpochMilli(0);  // very old
-            for (Record healthRecord : healthRecords) {
-                if (healthRecord instanceof HeightRecord heightRecord) {
-                    Instant heightTime = heightRecord.getTime();
-                    if (heightTime.isAfter(newestHeightTime)) {
-                        newestHeightTime = heightTime;
-                        profileParams.setHeight((float) (heightRecord.getHeight().getMeters()));
-                    }
-                } else if (healthRecord instanceof WeightRecord weightRecord) {
-                    Instant weightTime = weightRecord.getTime();
-                    if (weightTime.isAfter(newestWeightTime)) {
-                        newestWeightTime = weightTime;
-                        profileParams.setWeight((float) (weightRecord.getWeight().getKilograms()));
-                    }
-                } else {
-                    Logger.get().e("Unexpected record type: %s", healthRecord.getClass().getCanonicalName());
-                }
-            }
-            if (profileParams.isEmpty()) {
-                Logger.get().d("No new profile data to send");
-                return Tasks.forResult(null);
-            } else {
-                return sendHCProfileParams(nextToken, profileParams).onSuccessTask(
-                    localBackgroundExecutor, (apiCallResult) -> Tasks.forResult(apiCallResult.getValue()));
-            }
-        }));
-    }
-
-    @NonNull
-    public Task<Boolean> checkPermissions() {
-        HealthConnectAvailability availability = clientWrapper.getAvailability().getValue();
-        if (availability != HealthConnectAvailability.INSTALLED) {
-            Logger.get().e("Health Connect permission check failed: not installed");
-            return Tasks.forResult(false);
-        }
-        try {
-            Boolean hasAllPermissions =
-                clientWrapper.hasAllPermissionsAsync(clientWrapper.getDefaultRequiredPermissions()).get();
-            if (!hasAllPermissions) {
-                Logger.get().e("Health Connect permission check failed: does not have required permissions");
-                return Tasks.forResult(false);
-            }
-        } catch (ExecutionException ex) {
-            Logger.get().e("Health Connect permission check execution failed: %s", ex.getMessage());
-            return Tasks.forException(ex);
-        } catch (InterruptedException ex) {
-            Logger.get().e("Health Connect permission check execution interrupted: %s", ex.getMessage());
-            return Tasks.forException(ex);
-        }
-        return Tasks.forResult(true);
-    }
-
-    @NonNull
-    private Task<HealthConnectRecords> readIntradayMetrics(Set<FitnessMetricsType> optionsMetricTypes) {
-        TaskCompletionSource<HealthConnectRecords> taskCompletionSource = new TaskCompletionSource<>();
-        try {
-            String token = apiClient.getStorage().get(INTRADAY_CHANGES_TOKEN_KEY);
-            HealthConnectRecords records;
-            if (token == null) {
-                records = clientWrapper.getInitialIntradayRecordsAsync(MAX_DAYS, optionsMetricTypes).get();
-            } else {
-                records = clientWrapper.getIntradayChangeRecordsAsync(token, optionsMetricTypes).get();
-            }
-            taskCompletionSource.trySetResult(records);
-        } catch (ExecutionException ex) {
-            Logger.get().e("Health Connect read intraday metrics execution failed: %s", ex.getMessage());
-            taskCompletionSource.trySetException(ex);
-        } catch (InterruptedException ex) {
-            Logger.get().e("Health Connect read intraday metrics execution interrupted: %s", ex.getMessage());
-            taskCompletionSource.trySetException(ex);
-        }
-        return taskCompletionSource.getTask();
-    }
-
-    @NonNull
-    private Task<HealthConnectRecords> readProfileMetrics(Set<FitnessMetricsType> optionsMetricTypes) {
-        TaskCompletionSource<HealthConnectRecords> taskCompletionSource = new TaskCompletionSource<>();
-        try {
-            String token = apiClient.getStorage().get(PROFILE_CHANGES_TOKEN_KEY);
-            HealthConnectRecords records;
-            if (token == null) {
-                records = clientWrapper.getInitialProfileRecordsAsync(MAX_DAYS, optionsMetricTypes).get();
-            } else {
-                records = clientWrapper.getProfileChangeRecordsAsync(token, optionsMetricTypes).get();
-            }
-            taskCompletionSource.trySetResult(records);
-        } catch (ExecutionException ex) {
-            Logger.get().e("Health Connect read profile metrics execution failed: %s", ex.getMessage());
-            taskCompletionSource.trySetException(ex);
-        } catch (InterruptedException ex) {
-            Logger.get().e("Health Connect read profile metrics execution interrupted: %s", ex.getMessage());
-            taskCompletionSource.trySetException(ex);
-        }
-        return taskCompletionSource.getTask();
-    }
-
-    @NonNull
-    private Task<ApiCallResult<Void>> sendHCUploadData(@NonNull String tokenKey,
-                                                       @NonNull String nextToken,
-                                                       @NonNull HCUploadData uploadData) {
-        final TaskCompletionSource<ApiCallResult<Void>> sendDataTaskCompletionSource = new TaskCompletionSource<>();
-        Logger.get().d("Uploading HC data: %s", uploadData);
-        activitySourcesService.uploadHealthConnectData(uploadData).enqueue((apiCall, result) -> {
-            if (result.isError()) {
-                Logger.get().d("Failed to send HC data: %s", result.getError().getMessage());
-                final GoogleFitActivitySourceExceptions.CommonException exception =
-                    new GoogleFitActivitySourceExceptions.CommonException("Failed to send data to the server",
-                        result.getError());
-                sendDataTaskCompletionSource.trySetException(exception);
-                return;
-            }
-            Logger.get().d("Succeeded to send HC data");
-            apiClient.getStorage().set(tokenKey, nextToken);
-            sendDataTaskCompletionSource.trySetResult(result);
-        });
-        return sendDataTaskCompletionSource.getTask();
-    }
-
-    @NonNull
-    private Task<ApiCallResult<Void>> sendHCProfileParams(@NonNull String nextToken,
-                                                          @NonNull HCSynchronizableProfileParams profileParams) {
-        final TaskCompletionSource<ApiCallResult<Void>> sendDataTaskCompletionSource = new TaskCompletionSource<>();
-        Logger.get().d("Updating HC profile: %s", profileParams);
-        activitySourcesService.updateProfileOnBehalfOfHealthConnect(profileParams).enqueue((apiCall, result) -> {
-            if (result.isError()) {
-                Logger.get().d("failed to send the profile data: %s", result.getError().getMessage());
-                final GoogleFitActivitySourceExceptions.CommonException exception =
-                    new GoogleFitActivitySourceExceptions.CommonException("Failed to send data to the server",
-                        result.getError());
-                sendDataTaskCompletionSource.trySetException(exception);
-                return;
-            }
-            Logger.get().d("succeeded to send the profile data");
-            apiClient.getStorage().set(PROFILE_CHANGES_TOKEN_KEY, nextToken);
-            sendDataTaskCompletionSource.trySetResult(result);
-        });
-        return sendDataTaskCompletionSource.getTask();
-    }
-}
