+++ android/activitysources/src/main/java/com/fjuul/sdk/activitysources/entities/internal/healthconnect/data/HCWeightDataPoint.java
@@ -1,4 +1,4 @@
-package com.fjuul.sdk.activitysources.entities.internal.healthconnect;
+package com.fjuul.sdk.activitysources.entities.internal.healthconnect.data;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
@@ -6,9 +6,9 @@ import java.util.Locale;
 
 import androidx.annotation.NonNull;
 
-public class HCHeightDataPoint extends HCScalarDataPoint<Float> {
-    public HCHeightDataPoint(@NonNull Float cm, @NonNull Date start, @NonNull String dataSource) {
-        super(cm, start, dataSource);
+public class HCWeightDataPoint extends HCScalarDataPoint<Float> {
+    public HCWeightDataPoint(@NonNull Float kg, @NonNull Date start, @NonNull String dataSource) {
+        super(kg, start, dataSource);
     }
 
     @NonNull
@@ -17,6 +17,6 @@ public class HCHeightDataPoint extends HCScalarDataPoint<Float> {
         final SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS", Locale.getDefault());
         final String startFormatted = dateFormat.format(start);
         return String.format(Locale
-            .getDefault(), "HCHeightDataPoint: cm %.2f, start %s, dataSource %s", value, startFormatted, dataSource);
+            .getDefault(), "HCHeightDataPoint: kg %.1f, start %s, dataSource %s", value, startFormatted, dataSource);
     }
 }
