+++ android/activitysources/src/main/java/com/fjuul/sdk/activitysources/http/services/ActivitySourcesService.java
@@ -10,8 +10,8 @@ import com.fjuul.sdk.activitysources.entities.ConnectionResult;
 import com.fjuul.sdk.activitysources.entities.TrackerConnection;
 import com.fjuul.sdk.activitysources.entities.internal.GFSynchronizableProfileParams;
 import com.fjuul.sdk.activitysources.entities.internal.GFUploadData;
-import com.fjuul.sdk.activitysources.entities.internal.HCSynchronizableProfileParams;
-import com.fjuul.sdk.activitysources.entities.internal.HCUploadData;
+import com.fjuul.sdk.activitysources.entities.internal.healthconnect.data.HCSynchronizableProfileParams;
+import com.fjuul.sdk.activitysources.entities.internal.healthconnect.data.HCUploadData;
 import com.fjuul.sdk.activitysources.exceptions.ActivitySourcesApiExceptions;
 import com.fjuul.sdk.activitysources.http.ActivitySourcesApiResponseTransformer;
 import com.fjuul.sdk.activitysources.http.apis.ActivitySourcesApi;
